Add files via upload
About the client:
"Every day, ExpressFood creates 2 main dishes and 2 desserts at its headquarters with the help of its experienced chefs. These 2 main dishes are stored in a cold setting and then given to delivery people who bike to the food to the residence that requested it...in less than 20 minutes!



On its website, ExpressFood allows clients to order one or multiple dishes and desserts. There are no fees for delivery. Menu items change every day.

Once a client submits their order, they see a page displaying if the order has gone out for delivery yet, as well as an estimate for their meal's arrival time."

This project includes:
UML Diagrams and database,containing a list of clients,a list of the different daily menu items,a list of delivery people and an order history for each client.